Version of SRGAN where I find Nan error
In messy state where I have cells all over that print stuff to help find the Nan problem. Error occurs in preprocessing when a high-res image has the same min and max, and so gets divided by 0. However,  a bit confused by the fact that the min and max have 2 values, not 1, for normalization, thought there should just be 1 value.